smol refactoring
